Update) admin controller, router
