+++ views/admin.ejs
@@ -56,13 +56,15 @@
     }
 </style>
 <script>
-// 대기실 관리
+// 대기방 관리 페이지
 function waitingRoom(){
     window.location.replace("/admin/waitingRoom")
 }
+// 게임방 관리 페이지
 function gameRoom(){
     window.location.replace("/admin/gameRoom")
 }
+// 유저 관리 페이지
 function users(){
     window.location.replace("/admin/users")
 }
