+++ views/users.ejs
@@ -98,7 +98,7 @@ function getUserList() {
                         })
                     };
 
-    // 코멘트 수정
+// 유저 포인트 수정
 function editPoint(Id) {
     let point = $("#point").val();
     $.ajax({
@@ -114,7 +114,7 @@ function editPoint(Id) {
             }
     })
 };
-
+// 유저 삭제하기
 function deleteUser(id){
     let deleteConfirm = confirm("정말로 삭제하시겠습니까?")
     if (deleteConfirm === true) {
