+++ views/waitingRoom.ejs
@@ -57,7 +57,7 @@ $(document).ready(function () {
     getWaitingRooms();
 });
 
-// 대기실 불러오기
+// 대기방 불러오기
 function getWaitingRooms() {
     $.ajax({
         type: "GET",
@@ -85,7 +85,7 @@ function getWaitingRooms() {
                             }
                         })
                     };
-                    
+// 대기방 삭제하기
 function deleteRoom(roomNum){
     let deleteConfirm = confirm("정말로 삭제하시겠습니까?")
     if (deleteConfirm === true) {
